+++ index.html
@@ -18,6 +18,11 @@
                 <li id="contact"><a href="#">Contact</a></li>
             </ul>
         </nav>
+        <nav class="time">
+            <div id="date" class="">
+                
+            </div>
+        </nav>
 
         <header>
             <h1 id="awesome-heading">All Awesome books</h1>
@@ -49,7 +54,7 @@
 
         <p class="copy-right">Copyright ...</p>
     </div>
-
+    <script type="module" src="./modules/date.js"></script>
     <script defer src="./index3.js"></script>
 </body>
 </html>
