+++ style.css
@@ -30,6 +30,12 @@
   padding: 1rem;
 }
 
+.time {
+  border: 1px solid blue;
+  margin-left: 30rem;
+  margin-top: -2rem;
+}
+
 header {
   text-align: center;
   font-family: 'comic Sans MS', Arial, Helvetica, sans-serif;
